refactor: split auth request helpers
