+++ src/stdlib/auth.rs
@@ -45,6 +45,7 @@ use totp_rs::{Algorithm as TotpAlgorithm, Secret, TOTP};
 
 mod cookies;
 mod providers;
+mod request_helpers;
 mod routes;
 mod sessions;
 mod storage;
@@ -59,6 +60,11 @@ use providers::{
     validate_provider_name, value_map_to_provider,
 };
 pub use providers::{provider_to_value, value_to_provider};
+pub use request_helpers::{
+    auth_challenge_to_value, get_auth_challenge_id_from_request, get_session_id_from_request,
+    session_to_value, user_to_value,
+};
+use request_helpers::{get_host_and_proto, get_user_from_request};
 pub use routes::{
     handle_auth_callback, handle_auth_index, handle_auth_logout, handle_auth_protect,
     handle_auth_start,
@@ -2263,208 +2269,6 @@ pub fn verify_totp_code(secret: &str, code: &str, email: &str) -> bool {
     match create_totp(secret, email, "NTNT") {
         Ok(totp) => totp.check_current(code).unwrap_or(false),
         Err(_) => false,
-    }
-}
-
-/// Get session ID from request cookies
-pub fn get_session_id_from_request(request: &Value) -> Option<String> {
-    let config = get_auth_config()?;
-    let cookie_name = &config.cookie_name;
-
-    if let Value::Map(req_map) = request {
-        if let Some(Value::Map(headers)) = req_map.get("headers") {
-            if let Some(Value::String(cookie_header)) = headers.get("cookie") {
-                for cookie in cookie_header.split(';') {
-                    let parts: Vec<&str> = cookie.trim().splitn(2, '=').collect();
-                    if parts.len() == 2 && parts[0] == cookie_name {
-                        let signed_token = parts[1];
-                        // Verify HMAC signature and extract session ID
-                        return verify_session_id(signed_token, &config.session_secret);
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-pub fn get_auth_challenge_id_from_request(request: &Value) -> Option<String> {
-    let config = get_auth_config()?;
-    let cookie_name = auth_challenge_cookie_name(&config).ok()?;
-
-    if let Value::Map(req_map) = request {
-        if let Some(Value::Map(headers)) = req_map.get("headers") {
-            if let Some(Value::String(cookie_header)) = headers.get("cookie") {
-                for cookie in cookie_header.split(';') {
-                    let parts: Vec<&str> = cookie.trim().splitn(2, '=').collect();
-                    if parts.len() == 2 && parts[0] == cookie_name {
-                        let signed_token = parts[1];
-                        return verify_session_id(signed_token, &config.session_secret);
-                    }
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/// Get user from request as HashMap (internal helper)
-fn get_user_from_request(request: &Value) -> Option<HashMap<String, Value>> {
-    let session_id = get_session_id_from_request(request)?;
-    let session = get_session_by_id(&session_id)?;
-    if let Value::Map(m) = user_to_value(&session) {
-        Some(m)
-    } else {
-        None
-    }
-}
-
-/// Convert Session to User Value
-pub fn user_to_value(session: &Session) -> Value {
-    let mut map = HashMap::new();
-    map.insert("id".to_string(), Value::String(session.user_id.clone()));
-    map.insert(
-        "provider".to_string(),
-        Value::String(session.provider.clone()),
-    );
-    if let Some(email) = &session.email {
-        map.insert("email".to_string(), Value::String(email.clone()));
-    }
-    if let Some(name) = &session.name {
-        map.insert("name".to_string(), Value::String(name.clone()));
-    }
-    if let Some(picture) = &session.picture {
-        map.insert("picture".to_string(), Value::String(picture.clone()));
-    }
-    map.insert(
-        "raw".to_string(),
-        Value::Map(json_string_to_value_map(&session.raw_json)),
-    );
-    // Include CSRF token so apps can embed it in forms: {{user.csrf_token}}
-    if !session.csrf_token.is_empty() {
-        map.insert(
-            "csrf_token".to_string(),
-            Value::String(session.csrf_token.clone()),
-        );
-    }
-    Value::Map(map)
-}
-
-/// Convert Session to full Session Value
-pub fn session_to_value(session: &Session) -> Value {
-    let mut map = HashMap::new();
-    map.insert("id".to_string(), Value::String(session.id.clone()));
-    map.insert("user".to_string(), user_to_value(session));
-    map.insert("created_at".to_string(), Value::Int(session.created_at));
-    map.insert("expires_at".to_string(), Value::Int(session.expires_at));
-    map.insert(
-        "data".to_string(),
-        Value::Map(json_string_to_value_map(&session.data_json)),
-    );
-    map.insert(
-        "csrf_token".to_string(),
-        Value::String(session.csrf_token.clone()),
-    );
-
-    // Include token info if available
-    if let Some(ref at) = session.access_token {
-        map.insert("access_token".to_string(), Value::String(at.clone()));
-    }
-    if let Some(ref rt) = session.refresh_token {
-        map.insert("has_refresh_token".to_string(), Value::Bool(true));
-        // Don't expose refresh token directly for security
-        let _ = rt;
-    }
-    if let Some(exp) = session.token_expires_at {
-        map.insert("token_expires_at".to_string(), Value::Int(exp));
-    }
-
-    Value::Map(map)
-}
-
-pub fn auth_challenge_to_value(challenge: &AuthChallenge) -> Value {
-    let mut map = HashMap::new();
-    let user_id = if challenge
-        .subject_id
-        .starts_with(&format!("{}:", challenge.provider))
-    {
-        challenge.subject_id.clone()
-    } else {
-        format!("{}:{}", challenge.provider, challenge.subject_id)
-    };
-
-    map.insert("id".to_string(), Value::String(challenge.id.clone()));
-    map.insert(
-        "subject_id".to_string(),
-        Value::String(challenge.subject_id.clone()),
-    );
-    map.insert("user_id".to_string(), Value::String(user_id));
-    map.insert(
-        "provider".to_string(),
-        Value::String(challenge.provider.clone()),
-    );
-    map.insert("kind".to_string(), Value::String(challenge.kind.clone()));
-    map.insert("created_at".to_string(), Value::Int(challenge.created_at));
-    map.insert("expires_at".to_string(), Value::Int(challenge.expires_at));
-    map.insert(
-        "data".to_string(),
-        Value::Map(json_string_to_value_map(&challenge.data_json)),
-    );
-
-    Value::Map(map)
-}
-
-// ============================================================================
-// SECTION 8: Route Handlers
-// ============================================================================
-
-/// Helper to get host and protocol from request
-fn get_host_and_proto(req: &Value) -> (String, String) {
-    if let Value::Map(req_map) = req {
-        let host = req_map
-            .get("headers")
-            .and_then(|h| {
-                if let Value::Map(headers) = h {
-                    headers.get("host").and_then(|v| {
-                        if let Value::String(s) = v {
-                            Some(s.clone())
-                        } else {
-                            None
-                        }
-                    })
-                } else {
-                    None
-                }
-            })
-            .unwrap_or_else(|| "localhost:8080".to_string());
-
-        let proto = req_map
-            .get("headers")
-            .and_then(|h| {
-                if let Value::Map(headers) = h {
-                    headers.get("x-forwarded-proto").and_then(|v| {
-                        if let Value::String(s) = v {
-                            Some(s.clone())
-                        } else {
-                            None
-                        }
-                    })
-                } else {
-                    None
-                }
-            })
-            .unwrap_or_else(|| {
-                if host.contains("localhost") || host.starts_with("127.") {
-                    "http".to_string()
-                } else {
-                    "https".to_string()
-                }
-            });
-
-        (host, proto)
-    } else {
-        ("localhost:8080".to_string(), "http".to_string())
     }
 }
